% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{get_area_path}
\alias{get_area_path}
\title{Get filesystem path for a bay area}
\usage{
get_area_path(area)
}
\arguments{
\item{area}{Character: bay name (case-insensitive). One of:
\code{"guanabara"}, \code{"sepetiba"}, \code{"paranagua"}, \code{"santos"}, \code{"baiatos"}.}
}
\value{
Character: full path to area directory.
}
\description{
Get filesystem path for a bay area
}
