% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{get_exec_name}
\alias{get_exec_name}
\title{Get prediction executable name for an area}
\usage{
get_exec_name(area)
}
\arguments{
\item{area}{Character: bay name (case-insensitive). One of:
\code{"guanabara"}, \code{"sepetiba"}, \code{"paranagua"}, \code{"santos"}, \code{"baiatos"}.}
}
\value{
Character: executable filename.
}
\description{
Get prediction executable name for an area
}
