% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\docType{data}
\name{siscorar_areas}
\alias{siscorar_areas}
\alias{SISCORAR_AREAS}
\title{Available SISCORAR bay areas}
\format{
An object of class \code{character} of length 5.
}
\usage{
siscorar_areas()

SISCORAR_AREAS
}
\value{
Character vector of area names.
}
\description{
\code{SISCORAR_AREAS} is a character vector of the 5 Brazilian coastal bays
supported by SISCORAR. \code{siscorar_areas()} returns the same vector.
}
\details{
The supported areas are:
\itemize{
\item \strong{guanabara}: Guanabara Bay, Rio de Janeiro (~290,000 nodes)
\item \strong{sepetiba}: Sepetiba Bay, Rio de Janeiro (~5,400 nodes)
\item \strong{paranagua}: Paranagua Bay, Parana (~8,700 nodes)
\item \strong{santos}: Santos Bay, Sao Paulo (~8,900 nodes)
\item \strong{baiatos}: Baia de Todos os Santos, Salvador (~6,300 nodes)
}
}
\examples{
siscorar_areas()
SISCORAR_AREAS
}
\keyword{datasets}
