% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsiscorar-package.R
\docType{package}
\name{rsiscorar-package}
\alias{rsiscorar}
\alias{rsiscorar-package}
\title{rsiscorar: Interface to the SISCORAR Tidal Current Prediction System}
\description{
Predicts ocean currents for 5 Brazilian coastal bays using tidal harmonic analysis with 13 major constituents from the Brazilian Navy's SISCORAR (Sistema de Correntes e Mares) system. Optionally wraps the official SISCORAR executables (via Wine) for full 142-constituent regular-grid output. Exports to GRIB2, NetCDF, and GeoJSON formats for marine navigation software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leoniedu/rsiscorar}
  \item Report bugs at \url{https://github.com/leoniedu/rsiscorar/issues}
}

}
\author{
\strong{Maintainer}: Eduardo Leoni \email{e.leoni@gmail.com} (\href{https://orcid.org/0000-0003-0955-5232}{ORCID})

}
\keyword{internal}
