% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsiscorar-package.R
\docType{package}
\name{rsiscorar-package}
\alias{rsiscorar}
\alias{rsiscorar-package}
\title{rsiscorar: Interface to the SISCORAR Tidal Current Prediction System}
\description{
Provides an R interface to the Brazilian Navy's SISCORAR (Sistema de Correntes e Mares) tidal current prediction system. Wraps the official SISCORAR prediction executables (via Wine on macOS/Linux) to predict ocean currents for 5 Brazilian coastal bays using 142 tidal harmonics. Reads binary prediction files, queries currents at specific coordinates and times, and exports to GRIB2 and NetCDF formats for marine navigation software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leoniedu/rsiscorar}
  \item Report bugs at \url{https://github.com/leoniedu/rsiscorar/issues}
}

}
\author{
\strong{Maintainer}: Eduardo Leoni \email{e.leoni@gmail.com} (\href{https://orcid.org/0000-0003-0955-5232}{ORCID})

}
\keyword{internal}
